🐛Fix: WebSocketProvider 위치 조정
